+++ robin/src/error.rs
@@ -2,7 +2,7 @@ use serde_json;
 use redis;
 use std::{error, fmt};
 use job::JobName;
-use queue_adapters::JobQueueError;
+use queue_adapters::{JobQueueError, JobQueueErrorInformation};
 
 /// The result type used throughout Robin.
 pub type RobinResult<T> = Result<T, Error>;
@@ -10,9 +10,6 @@ pub type RobinResult<T> = Result<T, Error>;
 /// The different types of errors that might happen.
 #[derive(Debug)]
 pub enum Error {
-    /// The job we got from the queue isn't known.
-    UnknownJob(JobName),
-
     /// The job failed to perform and might be retried.
     JobFailed(Box<error::Error>),
 
@@ -21,12 +18,6 @@ pub enum Error {
 
     /// Some serialization/deserialization failed
     SerdeError(serde_json::Error),
-
-    /// Something related to Redis failed.
-    RedisError(redis::RedisError),
-
-    /// A Redis error that isn't included in `redis::RedisError`
-    UnknownRedisError(String),
 }
 
 impl fmt::Display for Error {
@@ -38,12 +29,9 @@ impl fmt::Display for Error {
 impl error::Error for Error {
     fn description(&self) -> &str {
         match self {
-            &Error::UnknownJob(ref name) => &name.0,
             &Error::JobFailed(ref err) => err.description(),
             &Error::JobQueueError(ref err) => err.description(),
             &Error::SerdeError(ref err) => err.description(),
-            &Error::RedisError(ref err) => err.description(),
-            &Error::UnknownRedisError(ref msg) => msg,
         }
     }
 }
@@ -54,14 +42,8 @@ impl From<serde_json::Error> for Error {
     }
 }
 
-impl From<redis::RedisError> for Error {
-    fn from(error: redis::RedisError) -> Error {
-        Error::RedisError(error)
-    }
-}
-
-impl From<JobQueueError> for Error {
-    fn from(error: JobQueueError) -> Error {
-        Error::JobQueueError(error)
+impl From<Box<JobQueueErrorInformation>> for Error {
+    fn from(e: Box<JobQueueErrorInformation>) -> Error {
+        Error::JobQueueError(e)
     }
 }